初步完成下载

+++ library/src/main/java/com/jake/library/download/IDownloadListener.java
@@ -13,5 +13,5 @@ public interface IDownloadListener {
 
     void onFail(String url);
 
-    void onProgress(String url, int positionSize, int totalSize);
+    void onProgress(String url, long positionSize, long totalSize);
 }
